router: split the weighted cluster out
